chore: move Chart -> _Chart
